+++ mobile/android/android-components/components/feature/prompts/src/main/java/mozilla/components/feature/prompts/login/LoginExceptions.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package mozilla.components.feature.prompts
+package mozilla.components.feature.prompts.login
 
 /**
  * Interface to be implemented by a storage layer to exclude the save logins prompt from showing.